Rename GamblerBets to BetTimelineListView and relocate components to the bet timeline module across Android and iOS.
