+++ Android/app/src/main/java/com/felipearpa/tyche/MainActivity.kt
@@ -16,7 +16,7 @@ import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.rememberNavController
 import com.felipearpa.foundation.emptyString
 import com.felipearpa.tyche.core.JoinPoolUrlTemplateProvider
-import com.felipearpa.tyche.gamblerbets.gamblerBetsNavView
+import com.felipearpa.tyche.bet.timeline.betTimelineListView
 import com.felipearpa.tyche.home.HomeRoute
 import com.felipearpa.tyche.bet.match.matchBetsNavView
 import com.felipearpa.tyche.home.homeNavView
@@ -133,7 +133,7 @@ fun Outlet(
             initialRoute = initialRoute,
         )
 
-        gamblerBetsNavView(navController = navController)
+        betTimelineListView(navController = navController)
 
         matchBetsNavView(navController = navController)
