+++ Android/app/src/main/java/com/felipearpa/tyche/poolhome/PoolHomeNavViewProvider.kt
@@ -6,7 +6,7 @@ import androidx.navigation.compose.composable
 import androidx.navigation.toRoute
 import com.felipearpa.tyche.bet.isLive
 import com.felipearpa.tyche.bet.match.MatchBetListViewRoute
-import com.felipearpa.tyche.gamblerbets.GamblerBetsViewRoute
+import com.felipearpa.tyche.bet.timeline.BetTimelineListViewRoute
 import com.felipearpa.tyche.home.HomeRoute
 import com.felipearpa.tyche.pool.poolscore.PoolScoreListRoute
 
@@ -32,7 +32,7 @@ fun NavGraphBuilder.poolHomeNavView(
             onGamblerOpen = { _, tappedGamblerId, tappedGamblerUsername ->
                 if (tappedGamblerId != route.gamblerId) {
                     navController.navigate(
-                        route = GamblerBetsViewRoute(
+                        route = BetTimelineListViewRoute(
                             poolId = route.poolId,
                             gamblerId = tappedGamblerId,
                             gamblerUsername = tappedGamblerUsername,
