+++ Android/bet/src/main/java/com/felipearpa/tyche/bet/timeline/BetTimelineListViewRoute.kt
@@ -1,9 +1,9 @@
-package com.felipearpa.tyche.gamblerbets
+package com.felipearpa.tyche.bet.timeline
 
 import kotlinx.serialization.Serializable
 
 @Serializable
-data class GamblerBetsViewRoute(
+data class BetTimelineListViewRoute(
     val poolId: String,
     val gamblerId: String,
     val gamblerUsername: String,
